+++ graphics/opengl/r_opengl_api.h
@@ -126,24 +126,14 @@ namespace Fresa::Graphics::API
         return id;
     }
     
-    template <typename V, typename U, typename I,
-              std::enable_if_t<Reflection::is_reflectable<V> && Reflection::is_reflectable<U> && std::is_integral_v<I>, bool> = true>
-    InstancedBufferID registerInstancedBuffer(const GraphicsAPI &api, const std::vector<V> &vertices,
-                                              const std::vector<U> &instanced_data, const std::vector<I> &indices) {
+    template <typename V, typename U, std::enable_if_t<Reflection::is_reflectable<V> && Reflection::is_reflectable<U>, bool> = true>
+    InstancedBufferID registerInstancedBuffer(const GraphicsAPI &api, const std::vector<V> &vertices, const std::vector<U> &instanced_data, ui32 vao) {
         static InstancedBufferID id = 0;
         do id++;
         while (instanced_buffer_data.find(id) != instanced_buffer_data.end());
         
         instanced_buffer_data[id] = InstancedBufferData{};
         InstancedBufferData &data = instanced_buffer_data.at(id);
-        
-        auto [vb, vao] = GL::createVertexBuffer(api, vertices);
-        data.vertex_buffer = vb;
-        data.vao = vao;
-        
-        data.index_buffer = GL::createIndexBuffer(api, indices);
-        data.index_size = (ui32)indices.size();
-        data.index_bytes = (ui8)sizeof(I);
         
         //: Get only the instanced attributes with updated positions
         auto attributes = API::getAttributeDescriptions<V, U>();